made earlier parity related fxns comply with type of testing fxns

+++ src/fxn.rs
@@ -1,10 +1,10 @@
 #![allow(dead_code)]
 
-pub fn build_16bit_parity_lookup() -> Vec<u8> {
-    (0..=0xFFFF).map(|i| parity(i)).collect()
+pub fn build_16bit_parity_lookup() -> Vec<u16> {
+    (0..=0xFFFF).map(|i| parity(i) as u16).collect()
 }
 
-pub fn parity(n: usize) -> u8 {
+pub fn parity(n: u64) -> u64 {
     /*
     finds the parity of a n-bit word
     */
@@ -33,7 +33,7 @@ pub fn swap_bits(n: usize, idx1: u8, idx2: u8 ) -> usize {
     swapped
 }
 
-pub fn parity_lookup(n: u64, lookup: &Vec<u8>) -> u8 {
+pub fn parity_lookup(n: u64, lookup: &Vec<u16>) -> u64 {
     /*
     used a cached lookup of table containing
     the parity of all 16bit strings to find the
@@ -44,7 +44,7 @@ pub fn parity_lookup(n: u64, lookup: &Vec<u8>) -> u8 {
     let p2 = lookup[((n >> 16) & bitmask) as usize];
     let p3 = lookup[((n >> 32) & bitmask) as usize];
     let p4 = lookup[((n >> 48) & bitmask) as usize];
-    (p1 ^ p2 ^ p3 ^ p4)
+    (p1 ^ p2 ^ p3 ^ p4) as u64
 }
 
 pub fn reverse_bits(n: u64, lookup: &Vec<u16>) -> u64 {
